+++ firmware/font.h
@@ -1,3 +1,3 @@
 extern const unsigned char smallFont[95*5];
 extern const unsigned char* getFontIndex (unsigned char ch);
-extern int string2pixels (const char* txt, unsigned char* buffer, int bufferLength);
+extern short string2pixels (const char* txt, unsigned char* buffer, short bufferLength);
