+++ firmware/trim.h
@@ -1,6 +1,6 @@
 extern bit trim_mode;
-extern int output_trim[TOTAL_OUTPUT_CHANNELS];
-extern int stick_center[TOTAL_OUTPUT_CHANNELS];
+extern short output_trim[TOTAL_OUTPUT_CHANNELS];
+extern short stick_center[TOTAL_OUTPUT_CHANNELS];
 
 #define NO_EXCEPTIONS -1
 #define EXCEPT
